Move scene manager to gui service (was in busy app)
