+++ applications/services/gui/scene_manager.c
