+++ applications/services/gui/scene_manager.h
